app/requirements.txt
fastapi
uvicorn[standard]
django-cors-headers
sqlalchemy
redis
python-dotenv
jinja2
pymysql
cryptography